services/similarity/requirements.txt
fastapi
uvicorn
pika
httpx
numpy<2       # <--- THIS IS THE CRITICAL FIX
pillow
torch
torchvision
faiss-cpu